+++ kani-driver/src/list/mod.rs
@@ -1,0 +1,6 @@
+// Copyright Kani Contributors
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+// Implements the list subcommand logic
+
+pub mod collect_metadata;
+mod output;
